ReportMessegの関数名を修正

+++ 3DAction_Project/3DAction_Project/DirectX.h
@@ -1,6 +1,12 @@
 ﻿#pragma once
 
 // ==============================================
+// 関数群の説明
+// ==============================================
+//DirectXの初期化、後処理関数
+
+
+// ==============================================
 // 前方宣言
 // ==============================================
 // ウィンドウハンドルの前方宣言
@@ -17,16 +23,22 @@ struct IDXGISwapChain;
 // SRVやUAVなどは別で作成する
 // RTも最終描画用で１つしか持たない
 // ================================================
-namespace DirectX11 {
-	bool Init(unsigned short Width, unsigned short Height, HWND windowHandle); // DirectXの初期化
-	void Uninit();                                                 // DirectXの初期化
+namespace DirectX11 
+{
+	// DirectXの初期化
+	bool Init(unsigned int Width, unsigned int Height, HWND windowHandle);
+	// DirectXの後処理
+	void Uninit();
 
+	// デバッグ描画
 	void BeginDraw(); // 描画バッファと深度バッファの初期化処理
 	void EndDraw();   // 次のバッファに入れ替える処理
 
 	void DebugDraw(float time); // デバッグ用描画
 
-	namespace Get { // ゲッター関数
+	// ゲッター関数
+	namespace Get 
+	{
 		ID3D11Device* GetDevice();
 		ID3D11DeviceContext* GetContext();
 		IDXGISwapChain* GetSwapChain();
